fix selection on iOS! I think!
This was a ton of work from Celine and I, eventually the solution we
seem to have settled on involves capturing clicks with a block floating
on top of the textblock and then triggering a focus into the textblock
seperately. We also need to check if the textblock becomes hidden by the
keyboard coming up and then scroll if neccessary.