more shenanigans to make ios focusing play nice
Basically we use a custom input element everywhere we need inputs and on
ios it intercepts clicks and focuses it manually. It also selects all
the text, if the user wants to select a specific text they just click
again but as inputs are used for things like urls and color values this
behavior fits everywhere we use them currently.