add basic list handling!
You should be able to indent, outdent, create and delete list blocks.

We're storing the data heierarchicaly but we're rendering it as a flat
list in the UI. This makes all the next/previous block logic more or
less just work.

There are still lots of places in the app where we're assuming a card's
block/children relations are the only children, which is no longer the
case, we need to use the query to recursively get all the lists
children.

We also need to wire up the copy and paste logic to handle lists, as
well as the previews on cards.

Finally, we need to to add shortcuts for list item movement. That should
be realtively easy!