WIP, set a default home theme
This is a WIP because I think there should be a better way to do this,
by passing something down to ThemeProvider directly. It's just tricky
because it calls the useColorAttribute hooks and so needs to be nested
in something. The way it's implemented now is *fine*, but a lil clunky!